# Speed up converting composite units to strings (with some formatters)