--- a/docs/changes/units/17043.perf.rst
+++ b/docs/changes/units/17043.perf.rst
@@ -1,0 +1,2 @@
+Converting composite units to strings with the ``"cds"``, ``"fits"``,
+``"ogip"`` and ``"vounit"`` formatters is now at least twice as fast.