@mhvk 2024-09-19
Benchmark is also on `units`, but clearly unrelated (with very large spread in new result).

I guess a question is, is whether this is important enough to add a benchmark. I think it is OK not too, since just based on elegance/simplicity I would have approved this PR.